allow FLEET to run more than once

+++ custom_code/hooks.py
@@ -12,6 +12,7 @@ from FLEET.data import main_assess
 import threading
 
 logger = logging.getLogger(__name__)
+logger.setLevel(logging.INFO)
 
 
 def query_mars(objectId):
@@ -87,15 +88,20 @@ def run_fleet(target):
     main_assess(target.name, target.ra, target.dec, output_filename=output_filename,
                 catalog_filename='', lightcurve_filename='', ztf_filename='', image_filename='')
     if os.path.exists(output_filename):
+        logger.info(f'FLEET pipeline finished for {target}.')
         dp, created = DataProduct.objects.get_or_create(
             target=target,
-            data=File(open(output_filename, 'rb')),
             data_product_type='image_file',
             product_id=f'{target.name}_FLEET'
         )
+        if created:
+            dp.data = File(open(output_filename, 'rb'))
+            os.remove(output_filename)
+            logger.info(f'{output_filename} saved')
+        else:
+            os.replace(output_filename, dp.data.path)
+            logger.info(f'{output_filename} replaced')
         dp.save()
-        os.remove(output_filename)
-        logger.info(f'{output_filename} saved')
     else:
         logger.warning(f'FLEET pipeline failed. Is {target} in PS1 3π?')
 
@@ -118,5 +124,6 @@ def target_post_save(target, created):
         gaia_data, gaia_url = query_gaia(gaia_name)
         save_gaia_photometry(target, gaia_data, gaia_url)
 
-    fleet_thread = threading.Thread(target=run_fleet, args=[target])
-    fleet_thread.start()
+    if created:
+        fleet_thread = threading.Thread(target=run_fleet, args=[target])
+        fleet_thread.start()
